+++ public/scripts/Arguments/Shape.js
@@ -1,7 +1,26 @@
 const Shape = ['square', 'circle', 'box'];
 
+const sSquare = "square";
+const sCircle = "circle";
 
-function DetermineShape () {
-    console.log("SHAPES");
-    
+function DetermineArgument() {
+
+  // Run loop to check all arugments
+  for(args of UserArgument) {
+
+     //check the users shape
+    if (sSquare.includes(args)) {
+
+      console.log("Square")
+
+    } else if (sCircle.includes(args)) {
+
+      console.log("Circle")
+
+    } else {
+
+      alert('ERROR USER COMMAND NOT RECOGNIZED')
+      console.log('ERROR USER COMMAND NOT RECOGNIZED');
+    }
+  }
 }