+++ CHANGELOG.md
@@ -5,10 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- Get Automation working
+### Adding
+- Work on UI
+- Make Shapes Customizable 
+- Secure backend
+- Fix bugs
+- Clean up code
 
-## [0.1.4] - 2019-06-18
+## [0.2.0] - 2019-07-07
+### Added
+-Fixed Changelog
+
+- Automation working
+- Fixed bugs
+- Can now determine the users Arugments and commands
+
+## [0.1.4] - 2019-06-27
 ### Added
 - re-wrote checking of User commands
 - added input_old.js for OLD js code that was Deprecated
@@ -16,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Useless old code from input.js
 
-## [0.1.3] - 2019-06-18
+## [0.1.3] - 2019-06-22
 ### Added
 - Statements to determine what the user wants
 - Startup with color
@@ -27,7 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [0.1.2] - 2019-06-18
+## [0.1.2] - 2019-06-21
 ### Added
 - User Arguments
 - Logging of the Input, Commands and Arguments with color